# fix #2786