@qm012 2021-07-24
hi @appleboy  @thinkerou  I updated the test report. thanks.

The latest benchmarks:

branch| benchmark
:-: | :-: 
**`master`**  | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/233564098 
**`this PR`** | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/233564278

The benchstat report:
```shell
$ benchstat old.txt new.txt
name              old time/op    new time/op    delta
Gin_Param           61.4ns ± 1%    60.3ns ± 4%  -1.69%  (p=0.006 n=9+9)
Gin_Param5           117ns ± 1%     116ns ± 1%  -1.43%  (p=0.000 n=9+9)
Gin_Param20          305ns ± 2%     303ns ± 0%  -0.67%  (p=0.027 n=10+10)
Gin_ParamWrite       114ns ± 2%     114ns ± 1%    ~     (p=0.346 n=9+9)
Gin_GithubStatic    80.2ns ± 1%    79.0ns ± 2%  -1.55%  (p=0.029 n=10+10)
Gin_GithubParam      139ns ± 1%     133ns ± 2%  -4.12%  (p=0.000 n=10+10)
Gin_GithubAll       27.7µs ± 1%    26.2µs ± 0%  -5.45%  (p=0.000 n=10+10)
Gin_GPlusStatic     63.2ns ± 3%    60.6ns ± 0%  -4.20%  (p=0.000 n=10+7)
Gin_GPlusParam      94.3ns ± 3%    89.6ns ± 1%  -4.98%  (p=0.000 n=10+10)
Gin_GPlus2Params     119ns ± 2%     115ns ± 0%  -3.59%  (p=0.000 n=10+10)
Gin_GPlusAll        1.22µs ± 3%    1.15µs ± 1%  -6.14%  (p=0.000 n=10+10)
Gin_ParseStatic     59.7ns ± 1%    61.6ns ± 1%  +3.18%  (p=0.000 n=9+10)
Gin_ParseParam      68.6ns ± 3%    68.7ns ± 1%    ~     (p=0.661 n=10+9)
Gin_Parse2Params    83.4ns ± 0%    84.7ns ± 1%  +1.52%  (p=0.000 n=8+10)
Gin_ParseAll        1.99µs ± 2%    1.96µs ± 0%  -1.77%  (p=0.000 n=10+9)
Gin_StaticAll       19.4µs ± 1%    18.5µs ± 1%  -4.82%  (p=0.000 n=10+9)
```